Remove repeating reference URLs
